+++ services/madoc-ts/src/frontend/shared/capture-models/editor/components/DocumentEditor/DocumentEditor.stories.tsx
@@ -26,6 +26,7 @@ const Inner = () => {
     addField: a.addField,
     setSelector: a.setSelector,
     setAllowMultiple: a.setAllowMultiple,
+    setRequired: a.setRequired,
     setLabelledBy: a.setLabelledBy,
     setPluralLabel: a.setPluralLabel,
   }));
@@ -39,6 +40,7 @@ const Inner = () => {
         setDescription={actions.setDescription}
         setLabel={actions.setLabel}
         setAllowMultiple={actions.setAllowMultiple}
+        setRequired={actions.setRequired}
         setLabelledBy={actions.setLabelledBy}
         setPluralLabel={actions.setPluralLabel}
         popSubtree={actions.popSubtree}