+++ services/madoc-ts/src/frontend/shared/capture-models/editor/input-types/BorderField/index.ts
@@ -17,6 +17,7 @@ const specification: FieldSpecification<BorderFieldProps> = {
   Component: BorderField,
   defaultValue: getEmptyBorder(),
   allowMultiple: true,
+  required: false,
   defaultProps: {
     clearable: true,
   },