+++ services/madoc-ts/src/frontend/shared/capture-models/new/components/DefaultFieldInstance.tsx
@@ -16,9 +16,10 @@ export const DefaultFieldInstance: React.FC<{
   const [isHighlighted, setIsHighlighted] = useState(false);
   const { immutableFields = [] } = useSlotConfiguration();
   const immutable = immutableFields.indexOf(property) !== -1;
-  const [selector, { isBlockingForm: disableForm }] = useResolvedSelector(field);
+  const [selector, { isBlockingForm }] = useResolvedSelector(field);
+  const disableForm = isBlockingForm;
 
-    useEffect(() => {
+  useEffect(() => {
     if (selector) {
       if (isHighlighted) {
         return helper.highlight(selector.id);