+++ services/madoc-ts/src/frontend/shared/form/CompositeInput.tsx
@@ -7,6 +7,12 @@ const Container = styled.div`
   display: flex;
   align-items: center;
   min-height: 2.75em;
+  
+  //&[data-is-invalid='true'] {
+  //  :not(:focus-within) {
+  //    border: 1px solid red;
+  //  }
+  //}
 `;
 
 const inputReset = css`