+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added notification for rejected submission and added view for rejected in contributions panel (MAD-1315)
 - Ability to add custom nav items to the Global Menu block
 - Added option item description to autocomplete (NS-26)
+- Added required fields in capture model (MAD-211)
+- Added not found pages for canvas, manifest, project and collection (MAD-1123)
 
 ### Changed
 - Changed language on user dash from 'reviews' to 'review tasks' to differentiate between the two